requirements.txt
# Core dependencies (required for benchmarking)
numpy>=1.26,<2.0  # Pin to 1.x for PyTorch compatibility
scipy>=1.11
scikit-learn>=1.4
opencv-contrib-python>=4.8,<4.10
Pillow>=10.2
tqdm>=4.66
PyYAML>=6.0
pandas>=2.0
hdbscan>=0.8  # Hierarchical DBSCAN for clustering

# Optional: Deep learning methods (ResNet50, DINOv2, OpenCLIP)
torch>=2.0; platform_system != "Darwin" or platform_machine != "arm64"
torch>=2.0; platform_system == "Darwin" and platform_machine == "arm64"
torchvision>=0.15
open-clip-torch>=2.20  # For OpenCLIP method

# Optional: Visualization and analysis
matplotlib>=3.8
seaborn>=0.13
fivecentplots>=0.5.4  # Advanced box plots for multi-experiment comparison
Jinja2>=3.1  # HTML gallery generation for clustering

# Optional: Jupyter notebook analysis
jupyter>=1.0
notebook>=7.0

# Optional: Streamlit web application
streamlit>=1.28.0

# Optional: AI Agent dependencies
python-dotenv>=1.0.0  # For loading .env files
google-generativeai>=0.3.0  # For Gemini agent

# Optional: Specialized models
# deepface>=0.0.79  # For face detection/recognition
# insightface>=0.7.3  # Alternative face detection backend
# mediapipe>=0.10.0  # Alternative face detection backend

# Optional: Earth Mover's Distance (advanced)
# POT>=0.9.3
